[v0.91.5][refactor][csdlc] Split prompt-template values internals (Closes #3622)
